chore: commit working tree (sionna backend, candidate-miss-wait, systematic analysis, outputs)
Bundles the Sionna PHY-abstraction BLER backend with in-flight work:
candidate-miss-wait tracking, systematic-analysis updates and plots,
midband experiment outputs, configs, and .claude skill files.

+++ src/scheduling_sim/config.py
@@ -11,7 +11,7 @@ class TrafficConfig:
     count: int
     packet_bits: int
     pdb_ms: int | None
-    period_slots: int | None = None
+    period_slots: float | None = None
     burst_cycle_interval: int | None = None
     gbr_bps: float = 0.0
     arrival_mode: str = "single_burst"
